fix: console encoding in color schemes
