+++ src/Xenial.Tasty.Reporters.Console/Reporters/Reporters.Console.cs
@@ -18,22 +18,6 @@ namespace Xenial.Delicious.Reporters
     {
         public static ColorScheme Scheme = ColorScheme.Default;
 
-        static ConsoleReporter()
-            => SetupConsoleEncoding();
-
-        private static void SetupConsoleEncoding()
-        {
-            if (RuntimeInformation.IsOSPlatform(OSPlatform.Windows))
-            {
-                //Register additional code pages for windows
-                //cause we deal directly with process streams
-                Encoding.RegisterProvider(CodePagesEncodingProvider.Instance);
-            }
-
-            Console.OutputEncoding = Encoding.UTF8;
-            Console.InputEncoding = Encoding.UTF8;
-        }
-
         public static TastyScope RegisterConsoleReporter(this TastyScope scope)
             => scope.RegisterReporter(Report)
                     .RegisterReporter(ReportSummary);
@@ -51,7 +35,6 @@ namespace Xenial.Delicious.Reporters
             catch (IOException) { /* Handle is invalid */ }
             return fallBackSize;
         });
-
 
         public static Task ReportSummary(IEnumerable<TestCase> tests)
         {
