+++ src/Xenial.Tasty/Reporters/Reporters.ColorScheme.cs
@@ -1,4 +1,6 @@
 ﻿using System;
+using System.Runtime.InteropServices;
+using System.Text;
 
 using static Xenial.Delicious.FeatureDetection.FeatureDetector;
 
@@ -6,6 +8,22 @@ namespace Xenial.Delicious.Reporters
 {
     public class ColorScheme
     {
+        static ColorScheme()
+            => SetupConsoleEncoding();
+
+        private static void SetupConsoleEncoding()
+        {
+            if (RuntimeInformation.IsOSPlatform(OSPlatform.Windows))
+            {
+                //Register additional code pages for windows
+                //cause we deal directly with process streams
+                Encoding.RegisterProvider(CodePagesEncodingProvider.Instance);
+            }
+
+            Console.OutputEncoding = Encoding.UTF8;
+            Console.InputEncoding = Encoding.UTF8;
+        }
+
         public ConsoleColor DefaultColor = ConsoleColor.White;
         public ConsoleColor ErrorColor = ConsoleColor.Red;
         public ConsoleColor WarningColor = ConsoleColor.Yellow;
